updating gitignore and initializing oracle dependency
